+++ src/test/activation/activationService.unit.test.ts
@@ -5,21 +5,19 @@
 
 // tslint:disable:max-func-body-length
 
-import { expect } from 'chai';
 import { SemVer } from 'semver';
 import * as TypeMoq from 'typemoq';
 import { ConfigurationChangeEvent, Disposable } from 'vscode';
-import { ExtensionActivationService, LSNotSupportedDiagnosticService } from '../../client/activation/activationService';
+import { ExtensionActivationService } from '../../client/activation/activationService';
 import {
     ExtensionActivators, FolderVersionPair,
     IExtensionActivationService, IExtensionActivator,
     ILanguageServerCompatibilityService,
     ILanguageServerFolderService
 } from '../../client/activation/types';
-import { CommandOption, IDiagnosticsCommandFactory } from '../../client/application/diagnostics/commands/types';
-import { DiagnosticCodes } from '../../client/application/diagnostics/constants';
+import { IDiagnosticsCommandFactory } from '../../client/application/diagnostics/commands/types';
 import { DiagnosticCommandPromptHandlerServiceId, MessageCommandPrompt } from '../../client/application/diagnostics/promptHandler';
-import { DiagnosticScope, IDiagnostic, IDiagnosticCommand, IDiagnosticFilterService, IDiagnosticHandlerService, IDiagnosticsService } from '../../client/application/diagnostics/types';
+import { IDiagnosticFilterService, IDiagnosticHandlerService } from '../../client/application/diagnostics/types';
 import {
     IApplicationShell, ICommandManager,
     IWorkspaceService
@@ -34,7 +32,6 @@ import { IServiceContainer } from '../../client/ioc/types';
 suite('Activation - ActivationService', () => {
     [true, false].forEach(jediIsEnabled => {
         suite(`Jedi is ${jediIsEnabled ? 'enabled' : 'disabled'}`, () => {
-            let diagnosticService: IDiagnosticsService;
             let serviceContainer: TypeMoq.IMock<IServiceContainer>;
             let pythonSettings: TypeMoq.IMock<IPythonSettings>;
             let appShell: TypeMoq.IMock<IApplicationShell>;
@@ -79,9 +76,7 @@ suite('Activation - ActivationService', () => {
                 serviceContainer.setup(s => s.get(TypeMoq.It.isValue(IDiagnosticFilterService))).returns(() => filterService.object);
                 serviceContainer.setup(s => s.get(TypeMoq.It.isValue(IDiagnosticsCommandFactory))).returns(() => commandFactory.object);
                 serviceContainer.setup(s => s.get(TypeMoq.It.isValue(IDiagnosticHandlerService), TypeMoq.It.isValue(DiagnosticCommandPromptHandlerServiceId))).returns(() => messageHandler.object);
-
-                diagnosticService = new LSNotSupportedDiagnosticService(serviceContainer.object);
-            });
+             });
 
             async function testActivation(activationService: IExtensionActivationService, activator: TypeMoq.IMock<IExtensionActivator>, lsSupported: boolean = true) {
                 activator
@@ -110,86 +105,13 @@ suite('Activation - ActivationService', () => {
 
                 await testActivation(activationService, activator, true);
             });
-            test('LS is not supported - Test Activation', async () => {
+            test('LS is not supported', async () => {
                 lanagueServerSupportedService.setup(ls => ls.isSupported()).returns(() => Promise.resolve(false));
                 pythonSettings.setup(p => p.jediEnabled).returns(() => jediIsEnabled);
                 const activator = TypeMoq.Mock.ofType<IExtensionActivator>();
                 const activationService = new ExtensionActivationService(serviceContainer.object, lanagueServerSupportedService.object);
 
                 await testActivation(activationService, activator, false);
-            });
-
-            if (!jediIsEnabled){
-                test('LS is not supported - Should display two options in message displayed with 2 commands', async () => {
-                    lanagueServerSupportedService.setup(ls => ls.isSupported()).returns(() => Promise.resolve(false));
-                    pythonSettings.setup(p => p.jediEnabled).returns(() => jediIsEnabled);
-                    let options: MessageCommandPrompt | undefined;
-                    const diagnostic = TypeMoq.Mock.ofType<IDiagnostic>();
-                    diagnostic.setup(d => d.code)
-                        .returns(() => DiagnosticCodes.InvalidEnvironmentPathVariableDiagnostic)
-                        .verifiable(TypeMoq.Times.atLeastOnce());
-                    const launchBrowserCommand = TypeMoq.Mock.ofType<IDiagnosticCommand>();
-                    commandFactory.setup(f => f.createCommand(TypeMoq.It.isAny(),
-                        TypeMoq.It.isObjectWith<CommandOption<'launch', string>>({ type: 'launch' })))
-                        .returns(() => launchBrowserCommand.object)
-                        .verifiable(TypeMoq.Times.once());
-                    const alwaysIgnoreCommand = TypeMoq.Mock.ofType<IDiagnosticCommand>();
-                    commandFactory.setup(f => f.createCommand(TypeMoq.It.isAny(),
-                        TypeMoq.It.isObjectWith<CommandOption<'ignore', DiagnosticScope>>({ type: 'ignore', options: DiagnosticScope.Global })))
-                        .returns(() => alwaysIgnoreCommand.object)
-                        .verifiable(TypeMoq.Times.once());
-                    messageHandler.setup(m => m.handle(TypeMoq.It.isAny(), TypeMoq.It.isAny()))
-                        .callback((_, opts: MessageCommandPrompt) => options = opts)
-                        .verifiable(TypeMoq.Times.once());
-
-                    await diagnosticService.handle([diagnostic.object]);
-
-                    diagnostic.verifyAll();
-                    commandFactory.verifyAll();
-                    messageHandler.verifyAll();
-                    expect(options!.commandPrompts).to.be.lengthOf(2);
-                    expect(options!.commandPrompts[0].prompt).to.be.equal('More Info');
-                });
-                test('LS is not supported - Should not display a message if the diagnostic code has been ignored', async () => {
-                    const diagnostic = TypeMoq.Mock.ofType<IDiagnostic>();
-
-                    filterService.setup(f => f.shouldIgnoreDiagnostic(TypeMoq.It.isValue(DiagnosticCodes.LSNotSupportedDiagnostic)))
-                        .returns(() => Promise.resolve(true))
-                        .verifiable(TypeMoq.Times.once());
-                    diagnostic.setup(d => d.code)
-                        .returns(() => DiagnosticCodes.LSNotSupportedDiagnostic)
-                        .verifiable(TypeMoq.Times.atLeastOnce());
-                    commandFactory.setup(f => f.createCommand(TypeMoq.It.isAny(), TypeMoq.It.isAny()))
-                        .verifiable(TypeMoq.Times.never());
-                    messageHandler.setup(m => m.handle(TypeMoq.It.isAny(), TypeMoq.It.isAny()))
-                        .verifiable(TypeMoq.Times.never());
-
-                    await diagnosticService.handle([diagnostic.object]);
-
-                    filterService.verifyAll();
-                    diagnostic.verifyAll();
-                    commandFactory.verifyAll();
-                    messageHandler.verifyAll();
-                });
-            }
-
-            test('LSNotSupportedDiagnosticService can handle LSNotSupported diagnostics', async () => {
-                const diagnostic = TypeMoq.Mock.ofType<IDiagnostic>();
-                diagnostic.setup(d => d.code)
-                    .returns(() => DiagnosticCodes.LSNotSupportedDiagnostic)
-                    .verifiable(TypeMoq.Times.atLeastOnce());
-                const canHandle = await diagnosticService.canHandle(diagnostic.object);
-                expect(canHandle).to.be.equal(true, 'Invalid value');
-                diagnostic.verifyAll();
-            });
-            test('LSNotSupportedDiagnosticService can not handle non-LSNotSupported diagnostics', async () => {
-                const diagnostic = TypeMoq.Mock.ofType<IDiagnostic>();
-                diagnostic.setup(d => d.code)
-                    .returns(() => 'Something Else')
-                    .verifiable(TypeMoq.Times.atLeastOnce());
-                const canHandle = await diagnosticService.canHandle(diagnostic.object);
-                expect(canHandle).to.be.equal(false, 'Invalid value');
-                diagnostic.verifyAll();
             });
 
             test('Activatory must be activated', async () => {
